Add a check for parallel prepping with rt_uniresource
This is an easy mistake to make, so let's check for it and be
informative rather than traipsing into potential trouble.  Warn and
error out of anyone's code tries to utilize rt_uniresource in a
multithreaded context - it's not intended for that.  Per struct resource
docs we need one resource per thread of execution, and accordingly
rt_uniresource can only be good for single threaded applications.