Stub in entry for edit mode setting method
Right now, the way edit modes are selected for solid editing tends to be the
faceplate GUI in MGED, which sets various values via its function callback.  To
be able to test this without the GUI, we want a way to trigger that same logic
without going through the MGED selection system.

The most direct route looks like refactoring the logic of the *_ed() functions
to be callable directly rather than only being exposed via the menu item static
functions.