Update code to use the Manifold 3.0 API
GLM has been swapped out for the public domain linalg.h math routines, and the
mesh containers have been reworked.  Using the manifoldConfig.cmake targets as
well.

The largest select drawing test differences appear to be the facetized mesh
interacting slightly differently with the LoD code.  One possibility is the use
of the 64 bit mesh container has slightly altered where vertices end up, which
may result in different clamping in the LoD scheme.  The meshes themselves test
as a valid manifold.

Note that at the moment this version has some regressions in boolean evaluation
performance observed in the plate mode evaluation - GenericTwin mesh 7316.1.t0
is one example.  Previous bot extrude timing was a little over 5 seconds to
completion, but the current code is far slower.  One possibility is an internal
change in Manifold to higher precision.  There is a new SetTolerance parameter
which may be useful for this case, but as of now still digging into what needs
to be done.  May need to back up to the previous version if we can't resolve
this before release.