Remove src/bext submodule in favor of an explicitly stored SHA1
Also, no longer deleting the cloned srcs and bext build from the BRL-CAD build
dir unless an option enabling that is explicitly set.  We still need it for CI,
since space is at a premium there and we need to clear the build outputs, but
otherwise removing these makes it more difficult to debug any problems.