Collapse dbot into bot dump
Not tested yet, but in prinicple this collapses dbot down to options on the
main bot dump command.  Will need to go through and systematically check the
refactored code to make sure it works.  Probably should set up some kind of
basic regressions on this, since it is frequently used for fast BoT geometry
extraction from .g files.