Fix behavior of dbconcat when given '/' prefix.
Older codes would treat "/" as indicating no prefix, but updated code ended up
treating it as an active string and adding slashes to names (which is a Bad
Idea).

Catch this case in the new code and avoid the issue.  May actually want to do
more sanity scrubbing on the input strings, but for the moment keep it simple.