Various improvements to the gltf-g importer
Trying to test bot dump, this importer wasn't giving the expected results.
Asset Import's export as glTF seems to be writing out a separate .bin file for
the buffer, while the bot dump experiments are serializing it as part of the
glTF file.  The latter apparently was problematic, and it looked as if the
import using the .bin file wasn't quite right either.

While I'm here, add some basic option handling - help, verbosity, etc.