Fix vertex data type - glTF only supports the FLOAT vertex coordinate type
Welp. Tinygltf apparently implements a data type the standard doesn't actually
call out - DOUBLE vertices aren't a standard type, so we can't just dump our
vertex data straight into a glb/glTF file and have it work.

That's a real shame, since it means a glTF export is inherently lossy for .g
BoT data, but the standard is what it is - even if tinygltf based readers could
support it, others (assimp being the observed case) don't support it.
Accordingly, step the data down to floats when writing it out.