Avoid (if possible) returning bu_editor paths with spaces on Windows
The command evaluation path for various commands involves using
bu_argv_from_string to create an argv from a programmatically generated string,
and it turns out that means trying to use full paths to editors when the editor
paths have spaces is Not Good.

The better fix would be to change how editors and editor options are passed to
libged, but there seem to be multiple commands with options that accept edit
strings.  We'll have to look at where and how those are being used and possibly
do some deprectations.