Have MGED editit commands used ged struct
Avoid putting the paths through string packing and unpacking.  This may
be all the uses of the -E options on these commands, in which case we
could probably just remove those options from the commands and remove
the editstring param completely from editit, but will go ahead and
deprecate to be safe.