Shift view limit defines into bv/defines and rename
Both libdm and libged use these, and previously we didn't have a good place to
store them that both could access.  As a result, they got dupliated when the
two were decoupled.  Now, however, we have libbv - that's a logical place to
stash these and is used by both libdm and libged.