Try a different approach to Orca_hash_tbl removal
Rather than requiring the caller to dupliate the tree walk, supply a callback
slot that gets passed the rtip and tsp.  The current region can be pulled off
of the rtip list, since it was just inserted, and that lets the caller perform
this (previously hard-coded operation) in the callback in the same way.

This not only alleviates concerns about whether the second tree walk in the
previous approach would behave the same way as the initial pass, but also
greatly simplifying the code required from the user to implement the feature.