Update callback type to include struct region pointer
The prior approach using the libbu standard callback type required that we pull
the region pointer off the head list, which in turn required that the head list
be stable until we were done with the callback.  While that woudl function, it
expanded the length of time when an individual thread would be blocking.

Instead, pass the specific region pointer of interest directly so we don't need
to worry about looking it up.