Add an XY rotation test for tor editing.
Phew.  Would have liked ft_edit_xy to be able to handle the setup for rotate as
well as translate, but looking at what MGED has in place for rotation that's
too heavy a lift - can't afford to dive into such a deep sinkhole and lose
sight of the functab editing refactor.

It looks as if at least some of the rotation modes and matricies probably back
the full generalized power of the knob command, which (among other things)
allows for continually spinning the model in X, Y and Z directions and has lots
of modes for both edit and view manipulation, with edits often falling back to
view manipulations when the MGED settings don't indicate a particular type of
edit mode is active.  This logic is also senstivie to the "coords" setting,
with different setup math in play for the matrices in reponse to the various
settings.

In the fullness of time this should all be boiled down to some sort of
"rotation" module that can capture and apply the full capability of the knob
command as an API, but that will have to be for another day.  Some of the more
obscure/esoteric modes would have to be studied to determine what they even do,
and decoding the various paths completely doesn't seem to be necessary for the
"primitive specific editing to functab" migration task.  Since the rotation ops
boil down in the end to a rotation flag being set and a matrix appliation, as
long as we are able to prepare both e_inpara and non-e_inpara inputs that
should (in theory) let us test the various paths of rt_solid_edit_generic_srot
successfully without needing to go deeper.

Note however that ft_edit_xy does remain unable to do the full setup for
generic solid object rotations (as opposed to param specific cases).  For API
cleanliness and completeness this should be revisited at some point.  The setup
logic in this test for the final XY rotation is an extraction of ONE possible
interpretation of XY deltas for rotation, but it is by no means the only
answer.  Devs looking into the problem can't treat it as the "uniquely correct"
way to interpret such mouse delta info, even if it should happen to match
MGED's default behavior in testing.