Modularize NIRT option parsing
Both MGED and the command line nirt program need to parse NIRT options.  The
libged nirt command does some custom setup to both handle options specified by
the app and prepare appropriate inputs to launch its own command line nirt(1)
command.  The hope is that by making the primary nirt option parsing available
as a library call, it will be easier to make the libged "pass through" more
robust.

The underlying motivation is a use case recently encountered with a very large
number of objects drawn in an MGED scene via search -exec.  Attempting to run
NIRT on such a scene will result in a very large argv command line being
constructed for nirt(1) launching, which does not scale - particularly on
Windows.  (We have had a similar problem with rt in the past when command lines
got too long due to complex scene specification.)  Newer NIRT has the ability
to accept "draw" commands to add new geometry objects, so we no longer need to
specify them all on the command line, but to make that work we must also make
sure we are constructing the appropriate inputs in libged.  If we can do option
processing in libged the same way we do it in nirt(1) itself, that should
(theoretically) let us handle things more cleanly at the libged layer as well
(separating out scripts and actual command-line only options, for example.)