Add user pointer tbl to struct ged, remove rt_cmd entries from ged_drawables struct
Looks like only two uses of the rt_cmd entries in the ged_drawables struct, and
the entires have long been marked deprecated.  The preview command logic looks
like it could stand some cleanup, but that's orthogonal to the focus right now
so the idea is to preserve its existing behavior as-is, just without involving
ged_drawable.

This also adds a way to allow apps and commands to stash arbitrary data
pointers in the ged struct.  Rather than an individual void pointer, which is
more typically what we provide for this, use a bu_ptbl so apps can avoid the
need to do their own packing and unpacking of multiple pointers.  struct ged is
big anyway and is already set up and cleared by functions, so the
flexibility/convenience benefit wins out over a single void pointer.