Checkpoint work done so far.
This represents an intermediate state after several iterations working with thie code.  After understanding
better what is going on, I'm thinking the correct answer for what this is trying to do is actually to have
nirt(1) generate a plot file as an optional output that can be read back in to provide the graphics.

Not only would that allow for one shot per ray instead of three and clean out all the complex format setting
and parsing, but it would allow us to specify -e scripts on a GED prompt that include shots beyond just the
one baked into libged's nirt itself that could also generate plot lines.  Would need to also support color
setting via the nirt option parsing to properly support the colors in the query ray control panel, but
that shouldn't be particularly difficult.