Rework how nirt is plotting and sending commands.
This addresses a number of issues.  The initial motivation was to allow for
processing a large number of drawn objects by avoiding having to list them all
on the command line.  In the process of reworking how data was being passed to
nirt, the plotting logic was simplified by having nirt(1) generate a plot file
as an optional output that can be read back in to provide the graphics.  This
avoiding the need to shoot multiple rays in libged to collect different data
types, as well as eliminating the textual post processing previously used to
recover the data.

Because option processing was modularized and reworked, it should now be
possible for virtually all options that are supported by the command line
nirt(1) command to also be used on the MGED command prompt.  It is not possible
to specify an alternate .g file to use there, but the command line parsing of
options is now shared by both nirt(1) and nirt(n).  If a -e option is detected
then the default "s" command for ray shooting will be skipped, allowing users
to customize what operations they want executed.

The nirt command in MGED will now also respect the formatting specified by the
qray command (fmt r, fmt p, etc.).  Previously it appears to have been using
the default libged nirt formatting.  This means that any existing .mgedrc
files (which may define other formatting) will now be active again - users
may wish to clear any qray settings pertaining to formatting from their .mgedrc
files to avoid unexpected output formatting changes.

The arbs.g file is useful for testing plotting - for example, the following
view settings will shoot down multiple arbs and trigger multiple plotting
output behaviors:

center -800 600 320
ae 322 0