Add another filtering criteria for plate extrude decimation
In addition to capping surface area changes, also watch to make sure we are
reducing the triangle count by at least some minimum threshold.  Currently
defaulting to .2 on a 0-1 scale for triangle counts - i.e. remove at least 20
percent of the triangles - and one one hundredth of the mesh area.  In other
words, don't work with the reduced mesh unless it's getting rid of at least 20%
of the original mesh triangles for an area change penalty of less than 1%.

The idea is to only do this on meshes where we can clear a lot of
near-degenerate triangles by doing this without significantly impacting the
overall mesh.  To be even better we should also "lock" any vertices that are
on interior holes, since as matters stand we might crush tiny holes completely
out of the mesh, but that will require an enhancement to the decimation
algorithm.

By making these parameters we can allow the user to decide if they want to pay
the potential performance penalties inherent in high accuracy - at least with
current Manifold algorithms those can occasionally be *severe*. Seconds to
minutes performance hits have been seen for relatively small meshes (less than
600 faces) when their structure is particulary hard for Manifold to handle.