Remove more stray references to mintie and minpieces
Obsolete with new BoT raytracing code.  There WAS still one place in librt
referencing the LIBRT_BOT_MINTIE environment variable, but it was not being
used in the traditional way - it was instead being used to allow overriding of
the maximum number of primitives allowed in a node during the HLBVH
acceleration structure building process.

In normal usage this variable shouldn't be modified (couuld be just a
debugging/tuning feature left from development) so rename it to be more
descriptive of what it is actually being used for.  May remove it entirely
later if there's no compelling motivation to allow user run-time tuning of that
parameter.