allocate the initial hit structs.
we were reallocating an array of struct hits that were never allocated in the first place.   encountered crash during linux testing during DA_APPEND on the first realloc.  possible could still be something else, like thread_id collision on the hit_arrays_per_cpu, but that would be a caller mistake to correct.