Zero the count, since we're reusing the array
This had a really weird test failure manifestation where the libged drawing
test that tessellates moss.g failed its diff test.  Inspection showed that the
LIGHT solid had been strangely tessellated, and further inspection led to the
discovery that the "cleanup" pass had deemed most of the triangles in the mesh
to be "thin".  That test is raytracing based, which made sense of the fact that
the prior commit could impact facetization.  A lot of digging by myself and
Sean determined that the prior memory allocation commit had also removed a
counter zeroing that is necessary for the internal results array to be
successfully reused.  For performance reasons the array is not zeroed out, so
stale data made debugging "interesting".  Restored the zero assignment with a
comment explaining the zero assignment is a functional necessity.