Detect and return failure in "soft fail" draw tests
If we're doing a "soft fail" for the image comparison in the drawing tests,
provide an error return code.  This way we can collect the results and have
the final overall return be success or failure.

This, in turn, should let us do our actual tests in soft fail mode and let us
see the full results in CI printing, rather than immediatly failing on just
one.  Due to the nature of the drawing tests, if we need to be using
approximate matching rather than hard comparisons and we tweak something we may
need to adjust the threshold values.  Rather than forcing an iteration for each
test, we want to run all the tests once to get a sense of the necessary
changes.