Translate MGED's knob cmd (view portion) to libged
The knob command in MGED appears to expose the most powerful view manipulation
capabilities exposed anywhere in BRL-CAD's user tools.  It interleaves editing
and view manipulations, requiring careful study to determine how to untangle
those without changing functionality.

The original code uses lots of switch statements for the various cases.  To
reduce the verbosity and consolidate redundant calculation logic, the libged
version of the code uses pointers to array entries rather than having
individual x,y,z case logic.

Although we're starting with the view manipulation logic, the main motivator
here is to try and encapsulate the editing manipulation capabilities of knob in
a way that will let us put them behind the librt editing interface.  The first
unit test setup for editing with the tor primitive made it clear that we'll
need to tackle knob if we want to properly expose all of MGED's editing power
via the functab interface - the test was able to set up one of the XY
manipulation cases (mouse input simulation) but there are a multitude of
possibilities with knob and if we don't want to have to keep repeating the
setup patterns of the initial test we'll need to get a handle on what an API
for this capability should look like.

For now, MGED's f_knob logic is unchanged - because it is so central to MGED's
interactive interface behaviors, the plan is to only replace its core logic
with libged or lower library functionality after the individual pieces have
been tested in isolation.

Because both "view knob" and MGED's "knob" command operate on a bview's knob
data, they may both be used to interact with the same view.

Note that we deliberately are not incorporting the MGED command's adc (angle
distance cursor) features into this command.  The MGED version is calling other
commands for adc features in any case, so the functionality is already exposed
elsewhere.