Don't call mged_svbase in MGED's view cmd wrapper
mged_svbase always resets the absolution portion of the knob values, and we
don't always or even usually want that for libged view commands.  Since we
are reworking the knob values to be part of bview, individual libged commands
can now reset them themselves if that is something they want to do.  Just
add the part of the logic from mged_svbase we do want.

Ran into this experimenting with a "view knob" command - doing this makes it
impossible for that command to work as it should.  MGED's knob command runs
without calling mged_svbase - to work the same way, "view knob" must also be
able to do the same thing.