Move per-command knob processing to libbv
Studying MGED's f_knob, being able to reuse the logic for handling individual
knob commands looks like it will be useful.  Moreover, there's no data
dependency on libged for the core command processing, so it can be pushed down
cleanly.

Probably need to get knob_rot and knob_tran moved as well, but will do that in
a separate commit.