Fix what appears to be a copy/paste error from 1997
The if conditional here is checking X rather than Y - this doesn't fit the
pattern of the other wrapping logic.  Digging back in the history, it appears
to originate from a 1997 mode to f_knob() in SVN commit 14080 (Git SHA1
b3fc3be99).

The prior blame commit (SVN 12861, SHA1 e78e1b2252) does not seem to have an
equivalent, and the commit that introduced it  does not mix X/Y/Z indices in
any of the other wrapping checks.

Noticed this while working to collapse the switch statements for chgview into
more readable code.