Start simplifying logic to disentangle editing
As with the view logic, we need to collapse all the switch statements
for increased readability.  The MGED state-specific conditional
determinations we try to wrap into a single up-front test, so we can
branch into either view or editing modes.  This is where the breakout of
the bv_knobs command processing into per-command becomes critically
important, because the MGED knob state awareness appears to be per
arg/val pair, not global to all arguments to an individual knob command.
(NOT something we will duplicate in the library versions, of course, but
to preserve that behavior in MGED's version we need enough flexibility
from the lower level functions to handle it.