Replace MGED edit state struct with librt struct
The librt structure is now defines a superset MGED's active editing varibales,
modulo a couple things like the global MGED state variable.  Use that struct
rather than the MGED version and put the few that aren't present there in
mged_state.

This again should amount to a rename, since we are still using one global copy
of the struct for all operations, but it gets us closer to using transient edit
states and having editing logic live in librt.