Separate the setup and editing execution for edit rot
That was more tangled than it appeared at first glance - separate out the
mged_state aware parts of the rot matrix logic and the setup.

This logic has setup for both sed and oed - still need to get a good handle on
how the latter works and how it relates to the solid edit data structures.