Stage edit_tran setup code in a function
To push this one down to librt, we'll need to have the functab
get_solid_keypoint replacement become live.  That is probably viable - it
should be one of the safest code migrations - but need to check over the logic
again so for now just get as close to a ready-to-move function as we can
without making that switch.