Consolidate sca knob editing code into chgview.c
Also do a preliminary separation of input setup and execution calls.  This one
is interesting in that it is the first blocker so far where the BE_O defines
are active.  They are specific to object editing rather than solid editing, so
they have not (as yet) been considered for the librt refactor.