Invert solid flag, refer to matrix edit instead
Rather than solid and object edit, at the API level it may be better to
think about the two behaviors here as solid edit - i.e. modify primitive
parameters - and matrix edit - apply the operation to the model_changes
matrix rather than the primitive parameters themselves.

MGED's objedit code is capable of editing both individual solids and
comb instances (see mover.c)  In the former case, it simply applies the
matrix directly in the final save-to-disk step (which may or may not
succeed, depending on the primitive.)  Comb instances, of course, only
support the matrix operations since they have no parameters to apply the
matrix to.

In theory, an xyz scale operation via solid edit or object edit ought to
wind up producing the same result, and there may be other cases where
the individual operations in isolation are the same, but there are
definitely some operations that will only work as matrix edits and
cannot be reliably assigned to all primitives - for example, an x-only
scale on tor will not work.  Another consequence will be wireframes -
solid editing the parameters will allow the generation of a new, data
appropriate wireframe, but a matrix edit will only provide enough
information for the application to distort the pre-existing wireframe
generated from the original primitive.

This may mean that capturing both matrix and solid operation logic for
rot, tra and sca ops at the librt API level will be enough to give us
what we need to express the MGED editing operations.  As MGED is
currently implemented I think it would suffice, but a more consistent
approach would be to have the matrix mode immediately attempt to update
the rt_db_internal instance of the solid when a new matrix is written to
model_changes.  Rather than manipulating the drawing of the scene solids
in the application's refresh loop based on a flag, it would be better to
have editing specific data associated with the edit object that could be
manipulated at will - that would provide more flexibility.

If the "solid" being edited is a comb or comb instance (neither of those
- particularly the latter - is supported yet but we clearly will need
to) that will always work as long as the matrix satisfies the general
criteria for all BRL-CAD geometry matrices.  Updating the rt_db_internal
for a comb would amount to applying the matrices to the comb child (or
all children if we're editing a "whole" comb rather than an instance in
its tree.  The child wireframes associated with the edit object created
for the comb would then be altered the same way MGED uses
new_edit_matrix and do_zoom to alter their scene appearance - since the
matrix would never be written to the leaf solids in that case, there
would be no strong motivation to regenerate the wireframes (and it
wouldn't always be possible anyway.)

On the other hand, a matrix application to a solid could be immediately
tested.  If it would succeed, a new and improved wireframe could be
generated, just like for solid edits.  If it would fail, we can (and
probably should) reject the attempt to apply the matrix and report an
error.

Conceptually, that may eventually make a matrix edit just another
category of input for an edit, along with e_mvalid and e_inpara...