Remove model2objview calc from librt edit code
Studying this more confirms my initial impression that this likely doesn't make
sense to do at this level.  Instead, try to explain what might drive the use of
a model2objview style matix and document how to set it up in the header
comments.

Thinking about this, I'm starting to wonder if it might be possible and beneficial
to apply a principle component analysis to meshes when they are first loaded and
establish a database-level hash map of origin centered and PCA aligned meshes to
try and spot xpushed meshes for drawing data reuse...