Clarify the material_name based warning in rtweight
The previous message made it look as if there was an error in the pinewood
rtweight calculation - the material_name attribute is set and there are no
material objects in his database, which is what prompted the warning, but
material_id is set and valid so the calculation can still proceed successfully.

Adjusted the reporting to be conscious of whether we also have a material_id to
use.  Since material_name IS set the user still needs to know we're not using
it, but until material objects become more commonplace a more mild sounding
note would seem to be in order.