Pull in latest librt_edit librt changes
Aside from the removal of the PSCALE define, which we add in sedit.h, there
shouldn't be any impact on MGED from these changes.

Renamed private functions for brevity, set up the oedit matrix editing logic in
functions.  Need to figure out if MGED exposes oed manipulations via the
command line as well as the mouse...