Take another stab at setting up an rt_comb_tess method
Revisits the work from d6bb98eda833, but this time we use the comb_internal to
hold the dbip and name rather than the higher level rt_db_internal structure.

comb primitives have to know about dbip and their origin object name, but that
is due to the specifics of how combs are defined rather than a general property
of rt_db_internal entities.  Accordingly, make that awareness an "internal"
detail of the comb definition.

It doesn't get us away from the awkwardness that a comb rt_db_internal can't
simply be used as-is with a different dbip (most other objects can do that) but
that limitation is innate to how combs are defined.  This approach should let
us make better use of the comb functab interface for methods beyond just import
and export, while not disturbing the broader rt_db_internal ecosystem.

Right now this is unused - should eventually replace the nmg boolean version of
tess in the facetize command, but there are also other considerations for a
general purpose comb facetizer to think about.  From an API perspective, we
arguably could/should wrap the entirity of the facetize csg->bot/nmg logic
under this call, but to fit it into this API we would need to solve the problem
of converting NMG polygons to triangles, doing a Manifold boolean, and then
simplifying the output triangles from Manifold back into NMG polygons to
produce the expected result.  (Manifold is triangle only, and we know we need
its robustness for boolean evaluation - if we can solve the problem of
converting back from evaluation triangles to NMG polygons we can probably just
remove nmg_boolean completely.)

Manifold has some abilities to track things like face IDs that should be
helpful - we could, for example, assign NMG ids and keep intact any NMG
polygons that weren't involved with the boolean changes - but there is no
question we would also have to solve re-creating NMG polygons from coplanar
output triangles.  Even though the face ID abilities recently enhanced for
Blender will help, there is no getting around that data structure translation.