adjust to not rely on libbu threading.
if we're not using libbu, then bu_parallel_id() will be always zero and we'll get collisions.  instead utilize a C11 thread local data container, that way we're always collision-free.  this could still be improved (avoid dynamic allocation altogether), but should be a minimal adjustment step towards release testing if we can rely on the C11 construct.