Move the db open/create/close defines to db_instance.h
They were in db_io.h, which is more or less reasonable, but since the idea
isn't to create a raw db_i struct without using the open/create functions, and
db_close does the destruction operation let's put these with the struct
definition itself.

Looking at db_i, it should have a private internal where the non-public aspects
of the struct are stored.  Since we have open/create and close doing the work
of setup and teardown anyway, that should be practical.