Add some mechanisms to manage child drawing
These are inactive in current code, but the idea is to set things up to allow
an rt_edit scene obj to have children inserted, and use the edit obj's settings
to draw them without disturbing the state of the child objs.

We won't be able to leave the original scene objs completely untouched, since
the main drawing loop will still need to know not to draw objs currently being
managed by the rt_edit scene obj, but this should help minimize the disruption.

Need to also look into whether the logic in the new drawing path is reusing
vlists for scene objs that are multiple instances of the same solid - if not,
this mechanism or something like it will be needed there as well.