Rough in a principle component analysis subcommand for BoTs
Principle Component Analysis (PCA) lets us take an input mesh (or really, any
point set) and "fits" a coordinate system to the data such that the axis
vectors of that system "line up" with the data.  The command will print out
the calculate data for the mesh.

If given an option for an object name, it will also produce a version of the
input mesh that has its center at the the global model origin and rotate it
according to the PCA - intuitively, it will be "lined up" with the global XYZ
model space coordinate system.

One potential application of this might be an "unpush-like" functionality for
meshes which would allow us to compare and recognize when meshes are the same
except for translations and rotations in model space.  The immediate motivator
is testing to see if we could use this to help minimze the number of vlists we
need to generate to represent solids - a view object could reference the
at-the-origin aligned data and provide the inverse matrix to the one used here
to restore the actual model position for drawing.  However, to catch such cases
we might need to do a slightly fuzzy comparison of the mesh vertcies to allow
for floating point drift due to the math operations involved in xpushing.