Set up a test to find PCA BoT matches
This logic tests out a scheme to use the PCA alignment of meshes in a .g file
to identify meshes that are copies of the same geometry relocated in space.

This is potentially useful in an "unpush" type scenario in that it would allow
us to identify candidates that could be replaced with references to a single
BoT and a matrix.  Less invasively, if this can be done fast enough it could
serve as a pre-processing step for drawing codes that would allow for a single
in-memory copy of the wireframe data that could be reused to draw all matching
objects.