Don't hit the disk for every BoT/BoT check
It didn't matter much for the GenericTwin small case, but the grouping problem
is fairly nasty when it comes to time complexity, assuming most geometry in a
.g database is distinct.  As a database gets bigger, if we have no groups at
all that can be defined we will have to check every object against every other
object to prove there are no matches - the only efficiencies we can get are to
not repeat an already performed check (e.g. if we've done A->B don't do the
B->A comparison) and pulling any matches we DO find out of the active set
(since we already have their group, there's no further need to check them.)

However, the initial implementation left some I/O performance on the table we
can recover.  Although it's true that in the worst case we might have to do PCA
comparisons for everything against everything, in practice counts will almost
always give us a negative answer immediately. There is no need to hit the disk
over and over just to pull the vert and face counts from the BoT internal
struts, since the memory overhead of storing just that data is basically
insignificant.  Accordingly, just pull that data once up front for all objects
and store it in maps.  Vastly speeds up runtimes on larger files needing lots
and lots of comparisons, since the previous way was unnecessarily hitting the
disk over and over.