Add some thoughts about edit obj drawing
Probably need to shelve this for now - consolidating the PARAM and MATRIX edit
ops is very appealing, but the undoubtedly tricky changes to the various
drawing systems needed will be a distraction from the edit refactor and unit
testing.  We CAN make things work as-is, so should do that first and then
circle back to the bits that (badly) need a rethink.