Start setting up to create GObj classes
One of the complexities of the new state management used for qged's drawing
path is the large number of maps and hash keys.  If the memory overhead and
performance of using class containers to hold most of this instead is
reasonable, it should simplify some things - however, we need to test out the
performance.  Since the change is very invasive and will need to be implemented
in stages, set up a branch to hold the incremental work.