Add lint solid raytrace checking to bot repair
Ideally we don't want bot repair returning something that will fail a lint check,
so add by default a step that throws those tests at the candidate BoT output.

One downside to doing this is we lose repairs that DO produce a manifold, which
may still work for facetize boolean evaluation even if they're not exactly what
we want for solid raytracing purposes.  Hence we add a setting that will allow
code to disable lint check.  This will let facetize proceed with repairs that
would be rejected at the user level to prepare boolean inputs and allow the
facetize process to continue.  There is no guarantee the final facetize output
will always make lint happy, but that was true in any case.