Start looking at implications of shifting scene obj ownership
Working with the gen2 drawing code, I'm getting the sense that libbv's bview
container is trying to do too much.  Right now it holds the scene objects with
the view itself, and goes through a lot of gyrations to have shared scene
objects in non-independent views and have per-view geometry in shared objects
for LoD.  That felt clunky at the time, and revisiting it I think that was the
wrong way to go - we need to separate scene object management from the views,
whose job should be camera-only.

The original idea was to have sparser and denser wireframes for the same object
for different views, but thinking about it more that complexity doesn't really
pay for itself - we just want one copy of any drawing data per object, at the
highest density needed for any given view, and we can just reuse that for all
cameras since (by construction) the LoD setting would be as high or higher than
necessary for the view.  Simultaneously maintaining lower resolution
representations for other views actually consumes MORE memory, since we still
need need the highest-required version for another view anyway.  The only
theoretical gain would be less drawing in some of the windows, and barring a
HUGE disparity between different view scales for an enormous number of objects
the likelihood of our having enough savings available to make that complexity
pay for itself is vanishingly small.

If we refocus on the idea of one object - LoD or not - that is pre-baked by the
app to have the vlists we want in it based on the app's knowledge of the views
and objects in the scene,  we then just have the app supply the final list of
objects for the libbv/libdm routines to draw. The whole mess of trying to track
db and view objs at the libbv level can just go away - it ends up baking too
much app state into the library design.  The app should just hand us a list of
objects to draw and the view - a "scene" - and let the drawing logic do its
thing.

However, divorcing scene object lists from bview will be a job, since the idea
of bview maintaining the view object lists is present in a lot of places. We'll
also have to rework where things like the free_scene_objs and vlfree lists are
managed - probably the GEDP makes sense, but we'll have to set that up and
pass a bunch of things through calls that previously could just pull them
from the views.

This commit is thoroughly non-functional - it's just trying get a rough sense
of what the necessary changes are and how much work it will be.