Add a functab slot for populating scene object data.
Haven't migrated the primitive specific functions like BoT yet, but this
establishes a general API pattern for handling multiple modes via a primitive
specific functab interface.

It felt a little off having primitive specific logic for drawing in libged -
hopefully this will give us a way to push the per-type details down to the
primitive implementations.

In support of this, the object point sampling routine got moved from libanalyze
to librt (it is fundamental to the mode 5 drawing.)  For the moment the
libanalyze version is deprecated rather than being removed outright.